* replace all `ushort` with `byte` in entity fields and related types which are already being stored as `TINYINT UNSIGNED` in mysql @ cralwer & imagePipeline + methods `ExtensionMethods.RoundToByte()` @ shared @ c#

+++ c#/crawler/src/Db/Post/IPostWithAuthorExpGrade.cs
@@ -2,5 +2,5 @@ namespace tbm.Crawler.Db.Post;
 
 public interface IPostWithAuthorExpGrade
 {
-    [NotMapped] public ushort AuthorExpGrade { get; set; }
+    [NotMapped] public byte AuthorExpGrade { get; set; }
 }
