+++ c#/crawler/src/Db/Revision/AuthorExpGradeRevision.cs
@@ -2,5 +2,5 @@ namespace tbm.Crawler.Db.Revision;
 
 public class AuthorExpGradeRevision : AuthorRevision
 {
-    public ushort AuthorExpGrade { get; set; }
+    public byte AuthorExpGrade { get; set; }
 }
