+++ c#/crawler/src/Db/TiebaUser.cs
@@ -7,7 +7,7 @@ public class TiebaUser : ITimestampingEntity
     public string? DisplayName { get; set; }
     public string Portrait { get; set; } = "";
     public uint? PortraitUpdatedAt { get; set; }
-    public ushort? Gender { get; set; }
+    public byte? Gender { get; set; }
     public string? FansNickname { get; set; }
     public byte[]? Icon { get; set; }
     public string? IpGeolocation { get; set; }
